+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fix spurious 'Pika Backup crashed' messages
 - Change to explain checkpoint creation when aborting backups
 - Change to restart backup after SSH connection timeout
+- Change reconnection to be abortable and count down seconds remaining
 - Add background portal status API messages about running operations
 - Add ability to answer questions from borg process