+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix potential crash after deleting archives
 - Fix spurious 'Pika Backup crashed' messages
 - Fix size estimation for some remote repositories
+- Fix issues with scheduled monthly backups scheduled for the end of the month
 - Change secret service error messages to include specific instructions how to resolve the issue
 - Change to explain checkpoint creation when aborting backups
 - Change to restart backup after SSH connection timeout